Add applicant manager on the main module
