--- ScholarshipManagement/ApplicantManagement/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-from django.http import HttpResponse
-
-# Create your views here.
-def searchApplicant(request):
-    HttpResponse("Hola :D")
