+++ ScholarshipManagement/ScholarshipModule/models.py
@@ -18,6 +18,19 @@ class Scholarships(models.Model):
     type = models.IntegerField(blank=False)
     requirements = models.TextField(blank=True)
 
+class Applicant:
+    studentCode = models.CharField(
+        primary_key=True,max_length=11,unique=True)
+    ID = models.IntegerField(
+        primary_key=True, auto_created=True, serialize=True, unique=True)
+    name = models.CharField(max_length=20,blank=False)
+    lastName = models.CharField(max_length=20,blank=False)
+    school = models.CharField(max_length=30,blank=False)
+    career = models.CharField(max_length=30,blank=False)
+    semester = models.IntegerField(blank=True)
+    email = models.CharField(max_length=40,blank=True)
+    phone = models.CharField(max_length=10,blank=True)  
+
 
 #Inheritance from an abstract class
 class User(AbstractBaseUser):
