+++ ScholarshipManagement/ScholarshipModule/templates/HTML/searchApplicant.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href='https://fonts.googleapis.com/css?family=Noto Music' rel='stylesheet'>
+    <title>
+        Busqueda de Aplicantes
+    </title>
+</head>
+<body>
+    <form method='POST'>
+        <label for="fname">Nombre:</label>
+        <input type="text" id="fname" name="fname"><br><br>
+        <button id="search" class="button">
+            <h5>
+                Buscar
+            </h5>
+        </button>
+    </form>
+</body>
+</html>
